break-out CollectionsIndex view as dedicated component
