+++ src/components/Collections/index.js
@@ -1,0 +1,82 @@
+import React, { useState } from 'react';
+import {
+  StyleSheet, View, SafeAreaView, StatusBar, TouchableOpacity,
+} from 'react-native';
+import { shape } from 'prop-types';
+import {
+  Header, Right, Body, Title, Left,
+} from 'native-base';
+import { Feather } from '@expo/vector-icons'; // eslint-disable-line import/no-extraneous-dependencies
+
+import Colors from '../../constants/Colors';
+import CollectionRow from '../CollectionRow/CollectionRow';
+import CollectionsDialog, { COLLECTIONS_DIALOG_ACTIONS, CollectionsDialogText } from '../Dialog/CollectionsDialog';
+
+const CollectionsIndex = ({
+  navigation,
+  collections,
+}) => {
+  const [collectionsDialogText, setCollectionsDialogText] = useState(null);
+
+  const handleNewCollectionPress = () => {
+    setCollectionsDialogText(CollectionsDialogText[COLLECTIONS_DIALOG_ACTIONS.CREATE]);
+  };
+
+  return (
+    <SafeAreaView style={styles.safeAreaView}>
+      <StatusBar backgroundColor={Colors.primary} barStyle="dark-content" />
+      <Header style={styles.header}>
+        <Left />
+        <Body>
+          <Title style={styles.headerText}>Collections</Title>
+        </Body>
+        <Right>
+          <TouchableOpacity onPress={handleNewCollectionPress}>
+            <Feather name="plus-square" size={24} color={Colors.headerIcon} />
+          </TouchableOpacity>
+        </Right>
+      </Header>
+      <View style={styles.collectionRowContainer}>
+        {Object.entries(collections).map(([id, { label }]) => (
+          <CollectionRow
+            key={id}
+            collectionId={id}
+            label={label}
+            navigation={navigation}
+          />
+        ))}
+      </View>
+      {collectionsDialogText && (
+        <CollectionsDialog
+          collectionsDialogText={collectionsDialogText}
+          setCollectionsDialogText={setCollectionsDialogText}
+        />
+      )}
+    </SafeAreaView>
+  );
+};
+
+CollectionsIndex.propTypes = {
+  navigation: shape({}).isRequired,
+  collections: shape({}).isRequired,
+};
+
+export default CollectionsIndex;
+
+const styles = StyleSheet.create({
+  safeAreaView: {
+    flex: 1,
+    backgroundColor: 'white',
+  },
+  collectionRowContainer: {
+    alignItems: 'center',
+  },
+  header: {
+    backgroundColor: Colors.headerBackground,
+    height: 50,
+  },
+  headerText: {
+    color: 'black',
+    fontSize: 18,
+  },
+});
