+++ src/screens/CollectionsListScreen.js
@@ -1,59 +1,22 @@
-import React, { useState } from 'react';
+import React from 'react';
 import {
-  StyleSheet, View, SafeAreaView, StatusBar, TouchableOpacity,
+  StyleSheet,
 } from 'react-native';
 import { connect } from 'react-redux';
 import { shape } from 'prop-types';
-import {
-  Header, Right, Body, Title, Left,
-} from 'native-base';
-import { Feather } from '@expo/vector-icons'; // eslint-disable-line import/no-extraneous-dependencies
 
-import Colors from '../constants/Colors';
-import CollectionRow from '../components/CollectionRow/CollectionRow';
-import CollectionsDialog, { COLLECTIONS_DIALOG_ACTIONS, CollectionsDialogText } from '../components/Dialog/CollectionsDialog';
+import CollectionsIndex from '../components/Collections/index';
 
 const CollectionsListScreen = ({
   navigation,
   collections,
 }) => {
-  const [collectionsDialogText, setCollectionsDialogText] = useState(null);
-
-  const handleNewCollectionPress = () => {
-    setCollectionsDialogText(CollectionsDialogText[COLLECTIONS_DIALOG_ACTIONS.CREATE]);
-  };
-
+  console.info('navigation: ', navigation);
   return (
-    <SafeAreaView style={styles.safeAreaView}>
-      <StatusBar backgroundColor={Colors.primary} barStyle="dark-content" />
-      <Header style={styles.header}>
-        <Left />
-        <Body>
-          <Title style={styles.headerText}>Collections</Title>
-        </Body>
-        <Right>
-          <TouchableOpacity onPress={handleNewCollectionPress}>
-            <Feather name="plus-square" size={24} color={Colors.headerIcon} />
-          </TouchableOpacity>
-        </Right>
-      </Header>
-      <View style={styles.collectionRowContainer}>
-        {Object.entries(collections).map(([id, { label }]) => (
-          <CollectionRow
-            key={id}
-            collectionId={id}
-            label={label}
-            navigation={navigation}
-          />
-        ))}
-      </View>
-      {collectionsDialogText && (
-        <CollectionsDialog
-          collectionsDialogText={collectionsDialogText}
-          setCollectionsDialogText={setCollectionsDialogText}
-        />
-      )}
-    </SafeAreaView>
+    <CollectionsIndex
+      navigation={navigation}
+      collections={collections}
+    />
   );
 };
 
@@ -69,19 +32,5 @@ const mapStateToProps = (state) => ({
 export default connect(mapStateToProps, null)(CollectionsListScreen);
 
 const styles = StyleSheet.create({
-  safeAreaView: {
-    flex: 1,
-    backgroundColor: 'white',
-  },
-  collectionRowContainer: {
-    alignItems: 'center',
-  },
-  header: {
-    backgroundColor: Colors.headerBackground,
-    height: 50,
-  },
-  headerText: {
-    color: 'black',
-    fontSize: 18,
-  },
+
 });
